Move the tab close button to better handle overflows when dealing with very small tabs
